zsh: prevent errors at login

+++ ripgrep/zprofile.zsh
@@ -1,5 +1,5 @@
 #!/bin/zsh
 THISDIR=${0:a:h}
 [ ! -L $XDG_CONFIG_HOME ] && {
-  ln -s $THISDIR/.ripgreprc $XDG_CONFIG_HOME/
+  [ ! -L $XDG_CONFIG_HOME/.ripgreprc ] && ln -s $THISDIR/.ripgreprc $XDG_CONFIG_HOME/
 }
